requirements.txt
# django framework 2.2.11
Django==3.1.4

# rest framework
djangorestframework==3.12.2

# to support env variables
python-dotenv==0.14.0

psycopg2==2.8.6
